requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# MCP dependencies
mcp>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# NBD Data Parser dependencies
undetected-chromedriver>=3.5.0
selenium>=4.15.0
pandas>=2.0.0
pdfplumber>=0.9.0
PyPDF2>=3.0.0
fake-useragent>=1.4.0

# Telegram Bot dependencies
aiogram>=3.0.0
aiohttp>=3.9.0

# AI Service dependencies
google-generativeai>=0.3.0